move leftover modify functions to otu.modify
